added cast attribute to products model added prepareForValidation() function inside the product store and update requests change selling_price db data type from float to decimal

+++ backend/database/migrations/2025_09_18_053243_create_products_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                 ->nullOnDelete();
 
             $table->string('unit', 50)->nullable();
-            $table->float('selling_price', 8)->nullable();
+            $table->decimal('selling_price', 15, 2)->nullable();
 
             $table->string('image_name', 50)->nullable();
             $table->string('original_image_name', 50)->nullable();
